Created a new webdemo which works exactly the same as the old, except that it uses Google's jsapi with jQuery to take away all the XmlHttpRequest checks, making it more compatible with other browsers and putting the focus of the javascript on the library rather then the request and parsing of the files.
Also revised the parsing a bit to make it possible to add additional regex keys to parse the file list for site files.

So short:
Better compatibility
More extensible parsing
Less code
Same functionality

I decided to add this version of the webdemo parallel to the old demo, until we are sure that this is the better option, I will create some tickets for that.

git-svn-id: http://google-sgnodemapper.googlecode.com/svn/trunk@163 902e2b25-573d-0410-8f12-e1b9ea7b2021